Update README.md
Using switch Statements

Step 1

This program is similar to the initial code from Lab 10.1. Add additional case values to the switch statement so the program will be able to handle lowercase letters entered as grades in addition to uppercase letters.

Sample Run:

What grade did you earn in Programming I ?
a
YOU PASSED!
an A - excellent work!

Reminder: The char type is used for variables that hold a single character. When using character literals in a program (like 'A') the character must be enclosed in single quotes, not double quotes that are used with string literals.

Step 2

Add an additional switch statement that will also output "YOU PASSED!" for a grade of D or better. You should add a whole new switch statement (not just add code to the original switch statement) and should not remove the original switch statement. Use the sample run given below to model your output.

Sample Run:

What grade did you earn in Programming I ?
A
YOU PASSED!
an A - excellent work!